raedgabsi

+++ src/main/java/com/pi/farmease/dao/CartItemsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CartItemsRepository extends JpaRepository<CartItems,Long> {
-    List<CartItems> findAllByCartUserId(Long userId);
+    List<CartItems> findAllByCartUserId(Integer userId);
     void deleteByCartCartId(Long cartId);
 }
